Database additions

+++ view.py
@@ -1,4 +1,5 @@
 from flask import Flask, render_template, current_app, request, redirect, url_for
+import db_queries as myDB
 from hashlib import sha256
 from datetime import datetime
 from user import User
@@ -9,12 +10,11 @@ from book import Book
 def home():
     
     db = current_app.config["db"]
-    currentuser_id = current_app.config["currentuser"]
-    
+    currentuser_id = current_app.config["currentuser"]    
 
     if currentuser_id is not None:
         user = db.getUser(currentuser_id)
-        username = user.username
+        username = user[1]
         return redirect(url_for("books", username=username))
 
     return render_template("home.html")
@@ -27,7 +27,7 @@ def login():
 
     if currentuser_id is not None:
         user = db.getUser(currentuser_id)
-        username = user.username
+        username = user[1]
         return redirect(url_for("books", username=username))
 
     if request.method == "GET":
@@ -41,7 +41,7 @@ def login():
         user = db.getUser(user_id)
         pw_bytestring = password.encode()
         password_hash = sha256(pw_bytestring).hexdigest() 
-        if user.password_hash == password_hash:
+        if user[2] == password_hash:
             current_app.config["currentuser"] = user_id
             return redirect(url_for("books", username=username))
 
@@ -54,7 +54,7 @@ def signup():
     
     if currentuser_id is not None:
         user = db.getUser(currentuser_id)
-        username = user.username
+        username = user[1]
         return redirect(url_for("books", username=username))
 
     if request.method == "GET":
@@ -66,9 +66,7 @@ def signup():
     if username != "" and password != "" and db.getUserByUsername(username) is None:
         pw_bytestring = password.encode()
         password_hash = sha256(pw_bytestring).hexdigest()
-        creation_date = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        db.addUser(User(username, password_hash, creation_date))
-
+        db.addUser(username, password_hash)
     return home()
 
 def logout():
@@ -110,16 +108,13 @@ def add_book(username):
     user_rating = request.form.get("user_rating")
 
     user_id = db.getUserByUsername(username)
-    book = Book(author, release_year=release_year, language=language, no_pages=no_pages, isbn=isbn)
-    book_id = db.addBook(book)
-    content = Content(title, content_type="book", type_specific_id=book_id)
-    content_id = db.addContent(content)
-    db.addUserContent(UserContent(user_id, content_id, completion_status=completed, owned=owned, user_rating=user_rating))
-
-    print(owned)
-    print(completed)
     
+    book_id = db.addBook(author=author, release_year=release_year, language=language, no_pages=no_pages, isbn=isbn)
+    content_id = db.addContent(title, content_type="book", type_specific_id=book_id)
+    db.addUserContent(user_id, content_id, completion_status=completed, owned=owned, user_rating=user_rating)
+
     return redirect(url_for("books", username=username))
+
 
 def delete():
     if request.method == "GET":
@@ -132,44 +127,75 @@ def delete():
     uc = db.getUserContent(currentuser_id, content_id)
     
     if uc is None:
-        return home()
+        return home() # go back?
 
     # GENERALIZE THE FUNCTION FOR ALL TYPES OF CONTENT
     # WITH IF BLOCK (CONTENT_TYPE)
 
     content = db.getContent(content_id)
-    book_id = content.type_specific_id
+    content_type = content[1]
 
     db.deleteUserContent(currentuser_id, content_id)
     db.deleteContent(content_id)
-    db.deleteBook(book_id)
+    
+    if content_type == "book":
+        book_id = content[2]
+        db.deleteBook(book_id)
+        
+
+    # elif content_type == "movies":
+    #     # delete movie
+    # elif content_type == "series":
+    #     # delete series
 
     return home()
+
 
 def edit(content_id):
     if request.method == "GET":
         return home()
 
-    print("Editing: " + str(content_id))
-    print(type(content_id))
-
-    """ db = current_app.config["db"]
+    db = current_app.config["db"]
     currentuser_id = current_app.config["currentuser"]
 
-    content_id = int(request.form.get("delete"))
     uc = db.getUserContent(currentuser_id, content_id)
     
     if uc is None:
-        return home()
+        return home() # go back?
+
+    title = request.form.get("book_title").strip()
+
+    if title == "":
+        return home() # go back?
 
     # GENERALIZE THE FUNCTION FOR ALL TYPES OF CONTENT
     # WITH IF BLOCK (CONTENT_TYPE)
 
-    content = db.getContent(content_id)
-    book_id = content.type_specific_id
+    author = request.form.get("author").strip()
+    release_year = request.form.get("release_year")
+    language = request.form.get("language").strip()
+    no_pages = request.form.get("no_pages")
+    isbn = request.form.get("isbn").strip()
+    completed = request.form.get("completed")
+    owned = request.form.get("owned")
+    user_rating = request.form.get("user_rating")
 
-    db.deleteUserContent(currentuser_id, content_id)
-    db.deleteContent(content_id)
-    db.deleteBook(book_id) """
+
+    content = db.getContent(content_id)
+    content_type = content[1]
+
+    db.updateUserContent(currentuser_id, content_id, completion_status=completed, owned=owned, user_rating=user_rating)
+    db.updateContent(content_id, title)
+    
+    if content_type == "book":
+        book_id = content[2]
+        db.updateBook(book_id, author=author, release_year=release_year, language=language, no_pages=no_pages, isbn=isbn)
+        
+
+    # elif content_type == "movies":
+    #     # delete movie
+    # elif content_type == "series":
+    #     # delete series
 
     return home()
+
